requirements.txt
astroid==2.0.4
Django==2.1.7
django-extensions==2.1.6
djangorestframework==3.9.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.7
six==1.11.0
wrapt==1.10.11
